LPS-61420 Make ElementComparator public class
